Wrap D3D12GetDebugInterface
Wrapping D3D12GetDebugInterface nicely requires also wrapping the
D3D12Debug and IUnknown types. The wrapping is currently very manual and
requires a lot of copy-paste, but results in no overhead.

There is probably some more to be done, such as traits, lifetime
management, and macros, but this is a good start and some proof of concept
work.