Wrap ID3D12Device::CreateRootSignature
I don't like the way we ensure the translated root parameters and static
samplers live long enough to be used. We do that by storing a vector of
them on the D3D12RootSignatureDescription object, but that requires the
object to be mutable.

We should either switch to interior mutability or find some other way. I
think the latter is the better choice.